reverting back to a single python file for random
All units except mllib/linalg.py fails because of the `random`
collision issue. I tried adding `globs = globals().copy()` to `_test()`
inside `testmod` but it didn’t help.
