+++ python/run-tests
@@ -67,10 +67,10 @@ run_test "pyspark/mllib/_common.py"
 run_test "pyspark/mllib/classification.py"
 run_test "pyspark/mllib/clustering.py"
 run_test "pyspark/mllib/linalg.py"
+run_test "pyspark/mllib/random.py"
 run_test "pyspark/mllib/recommendation.py"
 run_test "pyspark/mllib/regression.py"
 run_test "pyspark/mllib/tests.py"
-run_test "pyspark/mllib/rand/RandomRDDGenerators.py"
 
 if [[ $FAILED == 0 ]]; then
     echo -en "\033[32m"  # Green
